Move shipment entitlements to the entities file.
